Add a multiple-line TSV file parser.
Previously, we encountered an issue where the parsing of files was incorrect.
This was due to the files being split by line, but commit messages were multi-line.
Which resulted in messages being split at incorrect places, instead of concatenating them together.

Now, the parser has been updated to support multiple lines. This was resolved with the help of quotes.
When a file has multi line, the text is between quotes. When it's single-line, it doesn't have quotes.
The parser takes advantage of this, and resolves the previous issue.