Change character removal logic on the TSV parser.
Previously, the parser would remove 3 characters at the end of the line every time.

Now this has been changed to 2 characters and only if it's end of line.
This was because the parser stripped the characters at the final cycle, even if the line was not multiline.

After this change it parses the quote removal correctly.correctly.