changes to OpenAI section of app
the app did not work well with csv files with multiple lines. And
stripping /n characters caused the gpt-3 model to return wrong answers
there was no reason for .tsv files as export so that is changed to .csv files
the app can now work well with import multi line csv import and export.